backend/requirements/test.txt
-r base.txt

pycodestyle==2.5.0
mypy==0.770
isort==4.3.21
bandit==1.6.2
safety==1.8.5
django-stubs==1.5.0